requirements.txt
Django==1.10.4
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
psycopg2==2.6.2
PyJWT==1.4.2
